:zap:perf:Calculator step 4
