+++ 2024/24_11_14/Calculator/src/main/java/org/example/AbstractOperation.java
@@ -1,0 +1,7 @@
+package org.example;
+
+public abstract class AbstractOperation{
+    public double operate(int firstNumber, int secondNumber) {
+        return 0;
+    }
+}
