+++ 2024/24_11_14/Calculator/src/main/java/org/example/AddOperation.java
@@ -1,8 +1,8 @@
 package org.example;
 
-public class AddOperation{
-    public int operate(int firstNumber,int secondNumber){
+public class AddOperation extends AbstractOperation{
+    @Override
+    public double operate(int firstNumber, int secondNumber) {
         return firstNumber + secondNumber;
     }
-
 }
