+++ 2024/24_11_14/Calculator/src/main/java/org/example/DivideOperation.java
@@ -1,9 +1,8 @@
 package org.example;
 
-public class DivideOperation{
-    public double operate(int firstNumber,int secondNumber){
-        double v = (double) firstNumber / secondNumber;
-        return v;
-
+public class DivideOperation extends AbstractOperation{
+    @Override
+    public double operate(int firstNumber, int secondNumber) {
+        return (double)firstNumber / secondNumber;
     }
 }
