+++ 2024/24_11_14/Calculator/src/main/java/org/example/MultiplyOperation.java
@@ -1,7 +1,8 @@
 package org.example;
 
-public class MultiplyOperation{
-    public int operate(int firstNumber,int secondNumber){
+public class MultiplyOperation extends AbstractOperation{
+    @Override
+    public double operate(int firstNumber, int secondNumber) {
         return firstNumber * secondNumber;
     }
 }
